// src/types/enums/profile.enums.ts
export enum IStatementStatus {
  PENDING = 'PENDING',
  APPROVED = 'APPROVED',
  SUGGEST_CHANGES = 'SUGGEST_CHANGES',
}
export enum IStatementType {
  SUBMITTED = 'SUBMITTED',
  APPROVED = 'APPROVED',
  FEEDBACK = 'FEEDBACK',
}
export enum IndicatorStatus {
  IN_PROGRESS = 'IN_PROGRESS',
  PARTIALLY_MET = 'PARTIALLY_MET',
  UNMET = 'UNMET',
  MET = 'MET',
}
export enum Quarter {
  Q1 = 1,
  Q2 = 2,
  Q3 = 3,
  Q4 = 4,
}